prevent reusing a user uid
When a user is deleted, account is deleted. If later on someone
register with an old uid, depending on available apps, user may
gain access to old user jobs, data, etc.
To prevent this we record in old_users the uid of user when deleted
and at register with check that uid was not used.

Impact is if a legitimate user register again, he will not be able
to reuse his uid. However, before account validation, an admin
can modify his uid to reuse old one.