Add prevent_reuse and hash user id for reuser
In config file, adds general.prevent_reuse (bool) to prevent reusing
a user uid or group name once deleted. If prevent-reuse is not set in
config, default is true.

For user uids, use a md5 hash so that once user is deleted, no trace of
his *real* uid is kept in db (RGPD), system only needs to know it is
the same or different, not real value.