prevent reusing a user uid (#245)
* prevent reusing a user uid

When a user is deleted, account is deleted. If later on someone
register with an old uid, depending on available apps, user may
gain access to old user jobs, data, etc.
To prevent this we record in old_users the uid of user when deleted
and at register with check that uid was not used.

Impact is if a legitimate user register again, he will not be able
to reuse his uid. However, before account validation, an admin
can modify his uid to reuse old one.

* prevent reuse of group name

* fix init of old_groups

* Add prevent_reuse and hash user id for reuser

In config file, adds general.prevent_reuse (bool) to prevent reusing
a user uid or group name once deleted. If prevent-reuse is not set in
config, default is true.

For user uids, use a md5 hash so that once user is deleted, no trace of
his *real* uid is kept in db (RGPD), system only needs to know it is
the same or different, not real value.